+++ NEWS.md
@@ -1,3 +1,4 @@
+* New function `despecify_col()` removes specification decoration from a column.
 * Now using prepositions in specification strings.
 * Specification strings are now given in a package constant (`specify_notation`).
   These constants are referred throughout the package.